requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
azure-core>=1.38.2
azure-storage-blob>=12.24.0
azure.identity>=1.25.2
requests>=2.25.1
azure-mgmt-compute>=33.0.0